Refactor processLargeArrayAsync
